feat: partition a concurrent build during its scan (#6205)
Stacked on #6146. Review that one first; this PR replaces its approach.

## What

A `partition_by` index built with `CREATE INDEX CONCURRENTLY` or
`REINDEX CONCURRENTLY` now cuts its partitions during the scan, the way
a plain build does. #6146's rewrite pass after the build goes away, and
the demux merge with it.

## Why

#6146 left the concurrent build unpartitioned and rewrote its segments
afterwards, so the index paid a full extra rewrite. The reason the build
itself was ruled out: its phase 2 re-reads rows by ctid after the scan,
and with writers still running, a chain's live tail can be a version the
build's snapshot never saw.

That is a property of the fetch, not of the build. Phase 1 of a
concurrent build is a plain MVCC scan under the registered snapshot, and
visibility to a registered snapshot does not change. So phase 2 can
reproduce phase 1's choice exactly by resolving each chain under that
same snapshot.

## How

`build_index` plans boundaries for a concurrent build too. `do_build`
registers the scan's snapshot, which the scan itself drops when it ends,
and `drain_partitioned` pushes it as active and fetches query-visible
instead of maintenance-visible. `HeapDocFetcher` already had both modes;
a plain build keeps the maintenance one, which must index every live
row.

The serial path takes the snapshot from `build_index` directly; the
parallel path reads it off the shared scan descriptor.

`routable_dims` moves to `build_partitioning.rs`, next to the sampler,
since the build is now its only caller. `writer/demux.rs` is deleted.

## Tests

`partitioned_build_concurrently` (pg_regress) is unchanged except for
its header: two `CREATE INDEX CONCURRENTLY` indexes record the same
split points, in the same plan shape, with the same row counts as
before. The build now produces what the rewrite pass produced.

The four validation `#[pg_test]`s move to `build.rs` with the code they
cover.

## Open

- No test covers a concurrent build racing real writers. The property
this leans on (a registered snapshot's visibility is stable) is what
`validate_index` already relies on, but the test would need a second
backend.
- `pg_test_single_segment_build_exact_doc_cap_multiple` fails about half
the time, on this branch and on #6146 alike. It panics inside the
clustering dependency (`hierarchical.rs:250`) on random vector data,
unrelated to this change.

---------
