+++ nix/pg_search.nix
@@ -65,7 +65,7 @@ buildPgrxExtension (finalAttrs: {
   # If maintainers forget to do so, Nix will throw an error message that begins
   # like this and then provides the correct new hash:
   # error: hash mismatch in fixed-output derivation '...'
-  cargoHash = "sha256-3GNntQ8FVRMnwcvitTjc5nML1CnyBUVAOjQo5jw6VqY=";
+  cargoHash = "sha256-Zg/gq7yZOcBIkqET80tIfa0RRDcZAA4eyx2+v+PiUxs=";
 
   inherit cargo-pgrx postgresql;
